Hangsuu

+++ insider/src/main/java/com/kh/insider/dto/ReplyDto.java
@@ -6,9 +6,10 @@ import lombok.Data;
 
 @Data
 public class ReplyDto {
+
 	private int replyNo;
 	private int replyOrigin;
-	private int replyMemberNo;
+	private long replyMemberNo;
 	private String replyContent;
 	private int replyParent;
 	private int replyLike;
